Optimizations to better handle long lines
This changes add some optimizations for the plugin to work fast
even when opening files that may contain lines with tens of thousands
of characters (like minified js or css that can reach 1 million chars),
causing the editor to stall.

Besides the optimization, added the `~` character to the url pattern.

Finally, this change was also applied to the colorpreview, colorpicker
and selectionhighlight plugins with positive performance results.